Consumer and Reader do not throw FaultedException as excepted.
The Consumer and Reader didn't throw FaultedException in two scenarios where it should.

Consumer
1:
After the initial setup of the DotPulsar client where it can't connect to the cluster. Creating a consumer and using Receive didn't produce a ConsumerFaultedException as it should.

2:
After the initial setup of the DotPulsar client where it can't connect to the cluster and is in a faulted state. Creating a consumer and using Receive didn't produce a ConsumerFaultedException as it should.

Reader
1:
After the initial setup of the DotPulsar client where it can't connect to the cluster. Creating a reader and using Receive didn't produce a ReaderFaultedException as it should.

2:
After the initial setup of the DotPulsar client where it can't connect to the cluster and is in a faulted state. Creating a reader and using Receive didn't produce a ReaderFaultedException as it should.

All the scenarios are fixed, and tests are written in ConsumerTests.cs ReaderTests.cs confirming the fix.