Streamlined the release process for DotPulsar. (#186)
The version tag has been removed from the DotPulsar.csproj file, as the continuous delivery (CD) pipeline now uses the git tag as version information. Additionally, the ContinuousIntegrationBuild in the file has been removed and moved to publish-to-nuget.yml, so it can be handled by a command line switch to dotnet. We have also updated the release-process.md file to reflect the change that version information is no longer needed in the DotPulsar.csproj file.

Co-authored-by: David Jensen <djn@danskecommodities.com>