Update ToxiProxy and Pulsar images and adjust timeout intervals
The commit updates the used images of ToxiProxy and Pulsar in tests from versions 2.7.0 and 3.1.3 to 2.9.0 and 3.2.3 respectively to keep up with the latest versions. Also, it introduces longer timeout intervals in the PulsarClientTests, changing them from 10 seconds to 2 hours for more reliable results. Finally, it improves task cancellation functionality in ConnectionPool for more robust error handling.