Update CodeQL workflow to enhance security analysis (#220)
The CodeQL analysis workflow has been updated to include more detailed specifications for the analysis and build process. This includes adding information for runner and hardware resources considerations, permissions required for private CodeQL packs and private repositories, and increased detail on language analysis support. The workflow now utilizes dynamic runners based on language type and added support for the Swift language. The build mode is now customizable, which allows for more control over how the codebase is analyzed. The upgrade of the analysis version to v3 has resulted in overall improvements to the security analysis of the codebase.

Co-authored-by: David Jensen <djn@danskecommodities.com>